Enable two ways of arranging glyphs for a bit
Problem is that CGShowGlyphsAtPoint(cgcontext, x, y, glyphs) will take the origin of the glyph
into account, resulting in the glyph not actually being drawn at x,y.

This was mostly fine (it does cause a bug where the semicolon glyph
slightly intersects with a paren glyph cause the origins are not taken
into account) until I enabled ligatures.

With ligatures, some fonts (including Iosevka, the one I use), have
glyphs with very large origin offsets, for example the "!" in the "!="
glyph for Iosevka has a huge origin offset to the right so it intersects
nicely with the "=".

I thought the solution was simple: call CGShowGlyphsAtPoint(cgcontext, x
- glyph_rect.origin.x, y - glyph_rect.origin.y, glyphs) so the
the origin is at x,y

But this causes two problems:
1. Weird off-by-one bugs for the texture coordinates of glyphs, causing
the glyphs to look weird.

2. The cursor rendering is messed up now. The previous implementation
relied on the fact that the actual starting point of the glyph
is x + glyph_origin.x, which means that the glyph is in between
x and x < glyph_origin.x + glyph_origin.w <= glyph_advance.
Now x is glyph_origin.x

So I have both the old way and the new way up for now. I think these
problems should go away when I use CoreText to do the shaping.